requirements.txt
streamlit==1.31.0
groq==0.4.2
python-dotenv==1.0.1
pillow==10.2.0